mcp-server-ContextCraft/pyproject.toml
[project]
name = "mcp-server-contextcraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.6.0",
    "kiteconnect",
    "python-dotenv"
]
